requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
cryptography>=41.0.0
aiohttp>=3.9.0
cachetools>=5.3.0
requests>=2.31.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# Code Formatting (Development)
black>=23.11.0
